Move some utilities to super module to make them sharable
Makes it easier to use these types for different kinds of commit
walking. No matter how the history graph is traversed some concepts
remain identical.
